Move `FeedLoaderCacheImageDecorator` to production
