+++ EssentialApp/EssentialApp/FeedLoaderCacheImageDecorator.swift
@@ -1,0 +1,30 @@
+//
+//  FeedLoaderCacheImageDecorator.swift
+//  EssentialApp
+//
+//  Created by Jair Moreno Gaspar on 11/02/21.
+//
+
+import Foundation
+import EssentialFeed
+
+public class FeedLoaderCacheImageDecorator: FeedImageDataLoader {
+
+    private let loader: FeedImageDataLoader
+    private let cache: FeedImageCache
+    
+    public init(decoratee: FeedImageDataLoader, cache: FeedImageCache) {
+        self.loader = decoratee
+        self.cache = cache
+    }
+    
+    public func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        let task = loader.loadImageData(from: url, completion: { [weak self] result in
+            completion(result.map { data in
+                self?.cache.save(data, for: url) { _ in }
+                return data
+            })
+        })
+        return task
+    }
+}
