+++ EssentialApp/EssentialAppTests/FeedLoaderCacheImageDecoratorTests.swift
@@ -7,29 +7,7 @@
 
 import XCTest
 import EssentialFeed
-
-
-
-class FeedLoaderCacheImageDecorator: FeedImageDataLoader {
-
-    private let loader: FeedImageDataLoader
-    private let cache: FeedImageCache
-    
-    init(decoratee: FeedImageDataLoader, cache: FeedImageCache) {
-        self.loader = decoratee
-        self.cache = cache
-    }
-    
-    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
-        let task = loader.loadImageData(from: url, completion: { [weak self] result in
-            completion(result.map { data in
-                self?.cache.save(data, for: url) { _ in }
-                return data
-            })
-        })
-        return task
-    }
-}
+import EssentialApp
 
 class FeedLoaderCacheImageDecoratorTests: XCTestCase, FeedImageLoaderTestCase {
